delete me wip

+++ ckanext/validation/logic.py
@@ -664,25 +664,26 @@ def _run_sync_validation(resource_id, local_upload=False, new_resource=True):
 
         if not report['valid']:
 
-            # Delete validation object
-            t.get_action(u'resource_validation_delete')(
-                {u'ignore_auth': True},
-                {u'resource_id': resource_id}
-            )
-
-            # Delete uploaded file
-            if local_upload:
-                delete_local_uploaded_file(resource_id)
-
-            if new_resource:
-                # Delete resource
-                t.get_action(u'resource_delete')(
-                    {u'ignore_auth': True, 'user': None},
-                    {u'id': resource_id}
+            if not t.config.get('ckanext.validation.allow_invalid_data'):
+                # Delete validation object
+                t.get_action(u'resource_validation_delete')(
+                    {u'ignore_auth': True},
+                    {u'resource_id': resource_id}
                 )
 
-            raise t.ValidationError({
-                u'validation': [report]})
+                # Delete uploaded file
+                if local_upload:
+                    delete_local_uploaded_file(resource_id)
+
+                if new_resource:
+                    # Delete resource
+                    t.get_action(u'resource_delete')(
+                        {u'ignore_auth': True, 'user': None},
+                        {u'id': resource_id}
+                    )
+
+                raise t.ValidationError({
+                    u'validation': [report]})
     else:
         raise t.ValidationError({
             'validation': []
